### 2022-12-12 #119 @whedon
[PRE REVIEW]: Recommendation.jl: A Framework for Building Recommender Systems in Julia
**Submitting author:** @takuti (<a href="http://orcid.org/0000-0003-4561-4420">Takuya Kitazawa</a>)
**Repository:** https://github.com/takuti/Recommendation.jl
**Branch with paper.md** (empty if default branch):
**Version:**
