### 2022-12-12 #119 @takuti
@ranjanan Thank you for working on the matter. I have fixed the errors as shown above. Working PR is https://github.com/takuti/Recommendation.jl/pull/66 on the `paper-fix` branch - let me know if I should merge the branch or I can keep working on it throughout the review process.
