### 2022-12-15 #119 @takuti
@ranjanan If possible, I'd like to get some comments from [Adrian Salceanu](https://github.com/essenciary), the author of "Julia Programming Projects" using the Recommendation package, and [Abhijith Chandraprabhu](https://github.com/abhijithch) who I had a meaningful discussion about the package at JuliaCon 2019.
