### 2023-09-07 #119 @lucaferranti
@takuti , it seems some references have missing DOIs. While I look for reviewers, you may start filling those if you like. This could speed up the review process a bit later on.
