### 2023-09-07 #119 @lucaferranti
Hi @essenciary :wave: ,

would you be willing to review this JuliaCon submission? You can read more details about JuliaCon proceedings review process [here](https://juliacon.github.io/proceedings-guide/reviewer/) and you are of course welcome to ask me any clarifying questions you may have.
