+++ paper/ref.bib
@@ -5,7 +5,8 @@
   year         = 2017,
   pages        = { 65--98 },
   volume       = 59,
-  number       = 1
+  number       = 1,
+  doi          = { 10.1137/141000671 }
 }
 
 @inproceedings {ekstrand2020lenskit,
@@ -13,7 +14,8 @@
   title        = { {LensKit for Python: Next-Generation Software for Recommender Systems Experiments} },
   booktitle    = { {Proceedings of the 29th ACM International Conference on Information \& Knowledge Management} },
   year         = 2020,
-  pages        = { 2999--3006 }
+  pages        = { 2999--3006 },
+  doi          = { 10.1145/3340531.3412778 }
 }
 
 @inproceedings {gantner2011mymedialite,
@@ -47,7 +49,8 @@
   title        = { {Justifying Recommendations using Distantly-Labeled Reviews and Fine-Grained Aspects} },
   booktitle    = { {Proceedings of the 2019 Conference on Empirical Methods in Natural Language Processing and the 9th International Joint Conference on Natural Language Processing (EMNLP-IJCNLP)} },
   year         = 2019,
-  pages        = { 188--197 }
+  pages        = { 188--197 },
+  doi          = { 10.18653/v1/d19-1018 }
 }
 
 @inproceedings {Cantador:RecSys2011,
@@ -65,7 +68,8 @@
   pages        = { 61--70 },
   month        = December,
   volume       = 35,
-  number       = 12
+  number       = 12,
+  doi          = { 10.1145/138859.138867 }
 }
 
 @inproceedings {Sarwar2001,
@@ -83,7 +87,8 @@
   booktitle    = { Proceedings of SIGIR 1999 },
   year         = 1999,
   pages        = { 230--237 },
-  month        = August
+  month        = August,
+  doi          = { 10.1145/3130348.3130372 }
 }
 
 @article {Deshpande2004,
@@ -156,7 +161,8 @@
   booktitle    = { Recommender Systems Handbook },
   year         = 2011,
   chapter      = { 3 },
-  pages        = { 73--105 }
+  pages        = { 73--105 },
+  doi          = { 10.1007/978-0-387-85820-3\_3 }
 }
 
 @inproceedings {Geuens2015,
@@ -165,7 +171,8 @@
   booktitle    = { Proceedings of RecSys 2015 },
   year         = 2015,
   pages        = { 379--382 },
-  month        = September
+  month        = September,
+  doi          = { 10.1145/2792838.2796542 }
 }
 
 @article {Rendle2012-1,
@@ -175,7 +182,8 @@
   year         = 2012,
   month        = May,
   volume       = 3,
-  number       = 3
+  number       = 3,
+  doi          = { 10.1145/2168752.2168771 }
 }
 
 @inproceedings {Rendle2012-2,
@@ -184,7 +192,8 @@
   booktitle    = { Proceedings of WSDM 2012 },
   year         = 2012,
   pages        = { 133--142 },
-  month        = February
+  month        = February,
+  doi          = { 10.1145/2124295.2124313 }
 }
 
 @article {Rendle2012-3,
@@ -207,7 +216,8 @@
   title        = { {Evaluating Recommendation Systems} },
   booktitle    = { Recommender Systems Handbook },
   year         = 2011,
-  pages        = { 257--297 }
+  pages        = { 257--297 },
+  doi          = { 10.1007/978-0-387-85820-3\_8 }
 }
 
 @article {kotkov2016survey,
@@ -224,7 +234,8 @@
   title        = { {Improving Recommendation Lists through Topic Diversification} },
   booktitle    = { Proceedings of the 14th International Conference on World Wide Web },
[…]
