### 2023-09-10 #119 @takuti
Hi @lucaferranti, can you help me to know how MISSING DOIs are identified? I added `doi` fields to the reported citations, but it looks like the change doesn't resolve the warning.

https://github.com/takuti/Recommendation.jl/pull/66/commits/968df15cc924c80fe8353c436e596ad4b79f6657
