### 2023-09-11 #119 @lucaferranti
@takuti comparing this with other submissions where DOIs are fine, the only difference I see, is that others use curly braces instead of quotes, e.g. `doi = {10.1137/141000671}`, could you try that?

> I have fixed the errors as shown above. Working PR is https://github.com/takuti/Recommendation.jl/pull/66 on the paper-fix branch - let me know if I should merge the branch or I can keep working on it throughout the review process.

sorry I only now saw this never got answered. This depends on where you want to keep the `paper` folder, there are two strategies

1. Have the `paper/` in the main branch
2. Host the `paper/` in a separete branch

If you want the `paper/` to be in `main`, then I think you should merge the PR. If instead you dont want to incude the paer in the main branch, then you can close the PR and keep it as now. I see however you already have a `paper/` folder in main, so maybe you want to go with option 1?
