### 2023-09-12 #119 @takuti
@lucaferranti Thank you for the response. I replaced double-quotes with curly braces https://github.com/takuti/Recommendation.jl/pull/66/commits/082e38d3724d59fff13088123eab6d804c748252, but the warning still remains...

By the way, do we really need to "fix" the warning? It looks like the bot doesn't check the accuracy of DOIs, and JOSS documentation says these messages are just for reference.

Also thanks for commenting on the previous question about the branch. Yes, I'm taking Option 1, and I just wanted to check if I should merge the topic branch `paper-fix` to `main` NOW or AFTER peer-review is completed. It sounds like it doesn't really matter - I will merge once the DOI stuff is resolved.
