### 2023-09-13 #119 @lucaferranti
@takuti , the DOI issue should be fixed before final publication, but it does not of course prevent before starting the review (once I find reviewers). I will also try to take a closer look at the issue and try to help with that.
