### 2024-03-10 #119 @takuti
Hello @lucaferranti - Just checking in if I can still expect any updates for publication. If not in the near future, I will post the paper to arXiv not to waste my work for the time being. Let me know what your suggestion is.
