### 2024-03-10 #119 @lucaferranti
Hi, @takuti apologies for the delay! I will try to get a reviewer asap and hopefully this will move forward.

Regarding your question, arxiv is a preprint and you can post the paper there while it is under review. Hence, feel free to so so if you like.
