### 2024-04-09 #147 @lucaferranti
@PGimenez, @abhijithch, thank you very much for volunteering as reviewers!

I will be the editor for this submission, feel free to ping me to ask any questions you may have.

You can find review guidelines [here](https://juliacon.github.io/proceedings-guide/reviewer/) feel free to ask at any point if something is unclear.

As a first step, you can generate your checklist by running

You can write your review comments here or directly open an issue in the paper repository
