### 2024-04-09 #119 @xuanxu
@lucaferranti looks like it failed because there was no value for `version`
