### 2024-05-05 #147 @PGimenez
Hi, I'd like to submit my review of the paper "Recommendation.jl: A Framework for Building Recommender Systems in Julia" below.

I believe this is a good paper on recommender systems It does a deep enough review on the most popular methods used in the literature, and provides implementation details in the package. The paper is well-written, with clear language and good usage of math notation to explain concepts. I think the pape is **apt for publication with minor changes/corrections**.

There are a few minor points I'd like to raise, which can also be found in the attached annotated PDF:

 - It seems Figure 3 illustrates the fit and recommendation operations, not the data conversion into a DataAccessor

 - is it possible to create a DataAccessor from a matrix R?
R could indeed be converted into an event array by having the users and items be the row/column indices of the matrix, and the ratings be the entries of the matrix

- In 3.2.2 you mention that handling big matrices is very expensive. However, SVD is also known for being extremely costly ( O(N^3)) when used in matrix completion, as you mention in the next section 3.2.3. Therefore, SVD does not solve the resource cost problem.

- It is not clear to me how FMs relate to the (user,item) model used earlier, although I'll admit I hadn't seen FMs before.
I assume x would be a feature vector for a user, and y(x) would yield a rating for this vector. But, how is x related to the item being rated?

- Just before 4.2.1, Sentence seems incomplete, I'd reorder them like:
> let u be a target user, I the set of all items, and I_u the set of. truth items

Or leave as it is and add "be" before each variable like

> let a target user be u, the set of all items be I...

---

Then, a few more points:

**The cold-start problem**

In 3.4 you state that the cold-start problem arises when "there is not enough historical data to capture meaningful information". The question then is, how much data is needed? I believe a single rating is enough.

To me, the issue here is adding new users or items to the system since these haven't rated or haven't been rated by anyone yet. Therefore the system cannot find items similar to the one the user likes (which is none since they've rated nothing)

From a formulation standpoint, new users and items have an empty row/column in matrix R. When factorizing R with any of the earlier methods, the coefficients associated with the new user/item (p_i, q_j for instance) will be all-zero. This would yield no recommendations.

This can be solved by leveraging additional information in the form of feature vectors (the user `attribute` parameter) for either users or items. In fact, I believe this is what the content-based filtering in 3.4 is actually doing since it requires the user preferences (but no prior ratings).

This usage of attribute vectors can be incorporated into the matrix factorization technique as well, as shown in these papers

[Towards Open-World Recommendation: An Inductive Model-based Collaborative Filtering Approach
](https://www.cs.utexas.edu/users/inderjit/public_papers/imc_bioinformatics14.pdf)
[Matrix completion and extrapolation via kernel regression](https://core.ac.uk/download/pdf/286456343.pdf) (section 3. Disclosure, I wrote this one 😄)

**My point here is**: If most of the implemented algorithms can't solve the cold-start problem, how usable are they in real-world situations?

**Experimental results**

The results are good for showcasing the different implemented metrics and having a better understanding of their meaning. However, I find it lacking that only the SVD method is used. I'd expect the experiments to also show how this package allows to easily switch between algorithms. Also, some benchmarks as to their computational cost.

**Package documentation**

The package documentation is good, as it covers the algorithms in detail. It'd be good to also have one or two tutorials like the script included in the paper. (I there are already, sorry, I couldn't find any)
